requirements.txt
Django
django-push-notifications
django-rest-swagger
djangorestframework
